fix error

+++ Algebra/src/utils/token.ts
@@ -103,7 +103,7 @@ export function fetchTokenDecimals(tokenAddress: Address): BigInt {
   let decimalValue = BigInt.fromString("1")
   let decimalResult = contract.try_decimals()
   if (!decimalResult.reverted) {
-    decimalValue = BigInt.fromI32(decimalResult.value as i32) 
+    decimalValue = BigInt.fromI32(decimalResult.value) 
   } else {
     // try with the static definition
     let staticTokenDefinition = StaticTokenDefinition.fromAddress(tokenAddress)
